+++ src/main/java/com/backend/tienda/repository/ClientesRepository.java
@@ -3,5 +3,5 @@ package com.backend.tienda.repository;
 import com.backend.tienda.entidades.Clientes;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClientesRepository extends JpaRepository<Clientes, Double> {
+public interface ClientesRepository extends JpaRepository<Clientes, Long> {
 }
