+++ src/main/java/com/backend/tienda/servicios/UsuariosServicios.java
@@ -5,8 +5,8 @@ import java.util.List;
 
 public interface UsuariosServicios {
     public Usuarios actualizarUsuarios(Usuarios usuarios);
-    public void eliminarUsuarios(Double cedula_usuario);
+    public void eliminarUsuarios(Long cedula_usuario);
     public Usuarios guardarUsuarios(Usuarios usuarios);
     public List<Usuarios> listarUsuarios();
-    public Usuarios buscarUsuario(Double cedula_usuario);
+    public Usuarios buscarUsuario(Long cedula_usuario);
 }
